requirements.txt
# =====================
# Uwezo Capstone Project
# Backend + AI Processing
# =====================

# --- API / Web Framework ---
fastapi>=0.111.1
uvicorn>=0.23.2
pydantic>=2.7.1

# --- Database + ORM ---
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9

# --- PDF Handling / Extraction ---
pdfminer.six>=20231228
pdfplumber>=0.10.4
camelot-py[cv]>=1.0.9
reportlab>=4.1.0
pypdf<4.0,>=3.17


# --- Vision / OCR / IE ---
paddleocr>=2.7.0
layoutparser[ocr]>=0.3.4
opencv-python-headless>=4.12.0
pillow>=10.3.0
ultralytics>=8.3.0

# --- Data Processing / ML ---
torch>=2.2.0
transformers>=4.37.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.2
tqdm>=4.66.0

# --- Utilities ---
python-dotenv>=1.0.0
pyyaml>=6.0.0
psutil>=5.9.0
